+++ api/rs/slint/lib.rs
@@ -185,6 +185,7 @@ The follow table summarizes the entire mapping:
 | `relative-font-size` | `f32` | Relative font size factor that is multiplied with the `Window.default-font-size` and can be converted to a `length`. |
 | structure | `struct` of the same name | |
 | anonymous object | anonymous tuple | The fields are in alphabetical order. |
+| enumeration | `enum` of the same name | The values are converted to CamelCase |
 | array | [`ModelRc`] | Arrays are represented as models, so that their contents can change dynamically. |
 
 For user defined structures in the .slint, an extra struct is generated.